Rename log emitter to provider
