+++ opentelemetry-sdk/src/logs/mod.rs
@@ -1,8 +1,8 @@
 //! # OpenTelemetry Log SDK
 mod error;
 mod export;
-mod log_emitter;
 mod log_processor;
+mod logger_provider;
 pub(crate) mod record;
 
 /// In-Memory log exporter for testing purpose.
@@ -15,11 +15,11 @@ pub use in_memory_exporter::{InMemoryLogExporter, InMemoryLogExporterBuilder};
 
 pub use error::{LogError, LogResult};
 pub use export::{ExportResult, LogBatch, LogExporter};
-pub use log_emitter::{Builder, Logger, LoggerProvider};
 pub use log_processor::{
     BatchConfig, BatchConfigBuilder, BatchLogProcessor, BatchLogProcessorBuilder, LogProcessor,
     SimpleLogProcessor,
 };
+pub use logger_provider::{Builder, Logger, LoggerProvider};
 pub use record::{LogRecord, TraceContext};
 
 #[cfg(feature = "experimental_logs_batch_log_processor_with_async_runtime")]
